requirements.txt
websockets==12.0
websocket-client==1.7.0
confluent-kafka==2.3.0
asyncio==3.4.3
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
aiohttp==3.9.3
pandas==2.2.1
numpy==1.26.4
pyspark==3.5.1
requests==2.31.0 